@@ -1,15 +1,14 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 31
     defaultConfig {
         applicationId "com.yoti.mobile.android.sdk.yotidocscan.sample"
         minSdkVersion 21
-        targetSdkVersion 30
-        versionCode 270
-        versionName "2.7.0"
+        targetSdkVersion 31
+        versionCode 281
+        versionName "2.8.1"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
@@ -23,6 +22,10 @@ android {
         targetCompatibility = '1.8'
     }
 
+    buildFeatures {
+        viewBinding true
+    }
+
     buildTypes {
         release {
             minifyEnabled true
@@ -32,7 +35,7 @@ android {
                     'proguard-rules-retrofit.pro'
         }
         debug {
-             matchingFallbacks = ['release']
+            matchingFallbacks = ['release']
         }
     }
 }
@@ -40,11 +43,11 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'androidx.appcompat:appcompat:1.3.1'
-    implementation 'androidx.core:core-ktx:1.6.0'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'androidx.core:core-ktx:1.7.0'
 
-    implementation 'com.yoti.mobile.android.sdk:yoti-sdk-doc-scan:2.7.0'
-    implementation 'com.yoti.mobile.android.sdk:yoti-sdk-doc-scan-sup:2.7.0'
-    implementation 'com.yoti.mobile.android.sdk:yoti-sdk-liveness-zoom:2.7.0'
-    implementation 'com.yoti.mobile.android.sdk:yoti-sdk-facecapture:2.7.0'
+    implementation 'com.yoti.mobile.android.sdk:yoti-sdk-doc-scan:2.8.1'
+    implementation 'com.yoti.mobile.android.sdk:yoti-sdk-doc-scan-sup:2.8.1'
+    implementation 'com.yoti.mobile.android.sdk:yoti-sdk-liveness-zoom:2.8.1'
+    implementation 'com.yoti.mobile.android.sdk:yoti-sdk-facecapture:2.8.1'
 }
