@@ -9,7 +9,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import kotlinx.android.synthetic.main.bottom_sheet.*
+import com.yoti.mobile.android.sdk.yotidocscan.websample.databinding.BottomSheetBinding
 
 private const val SESSION_PREFERENCES_ID = "SESSION_PREFERENCES_ID"
 private const val SESSION_CONFIGURATION_KEY = "SESSION_CONFIGURATION_KEY"
@@ -19,6 +19,9 @@ class SessionBottomSheet: BottomSheetDialogFragment() {
     private var sessionConfigurationListener: SessionConfigurationListener? = null
     lateinit var sharedPreferences: SharedPreferences
 
+    private var _binding: BottomSheetBinding? = null
+    private val binding: BottomSheetBinding get() = _binding!!
+
     interface SessionConfigurationListener {
         fun onSessionConfigurationSuccess(sessionUrl: String)
         fun onSessionConfigurationDismiss()
@@ -26,23 +29,28 @@ class SessionBottomSheet: BottomSheetDialogFragment() {
 
     override fun onCreateView(
             inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
-    ): View? = inflater.inflate(R.layout.bottom_sheet, container, false)
+    ): View {
+        _binding = BottomSheetBinding.inflate(inflater, container, false)
+        return binding.root
+    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        session_url.setText(sharedPreferences.getString(SESSION_CONFIGURATION_KEY, ""))
+        with (binding) {
+            sessionUrl.setText(sharedPreferences.getString(SESSION_CONFIGURATION_KEY, ""))
 
-        start_session_button.setOnClickListener {
-            sessionConfigurationListener?.onSessionConfigurationSuccess(
-                    session_url.text.toString()
-            )
-            this.dismiss()
+            startSessionButton.setOnClickListener {
+                sessionConfigurationListener?.onSessionConfigurationSuccess(
+                        sessionUrl.text.toString()
+                )
+                dismiss()
+            }
         }
     }
 
     override fun onDismiss(dialog: DialogInterface) {
-        sharedPreferences.edit().putString(SESSION_CONFIGURATION_KEY, session_url.text.toString()).apply()
+        sharedPreferences.edit().putString(SESSION_CONFIGURATION_KEY, binding.sessionUrl.text.toString()).apply()
         sessionConfigurationListener?.onSessionConfigurationDismiss()
         super.onDismiss(dialog)
     }
@@ -51,6 +59,11 @@ class SessionBottomSheet: BottomSheetDialogFragment() {
         super.onAttach(context)
         sharedPreferences = requireContext().getSharedPreferences(SESSION_PREFERENCES_ID, Context.MODE_PRIVATE)
         sessionConfigurationListener = context as? SessionConfigurationListener
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
     }
 
     override fun onDetach() {
